feat(yard): moving fns out of yard.tsx
